requirements.txt
groq
python-dotenv
edge-tts
requests
moviepy==1.0.3
pillow
imageio
imageio-ffmpeg